pyproject.toml
[project]
name = "t2s"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "llvmlite>=0.44.0",
    "nltk>=3.9.1",
    "numba>=0.61.2",
    "openai-whisper",
    "rich>=14.0.0",
    "sounddevice>=0.5.2",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
